+++ projects/hx-ui/src/lib/tabular/tabular-config.interface.ts
@@ -1,5 +1,6 @@
 import {TabularSize} from './tabular-size.enum';
 import {ISortByProperty, SortByDirection} from './tabular-sort-by.service';
+import {TabularTheme} from './tabular-theme.enum';
 
 export abstract class ITabularConfig {
   id: string;
@@ -11,9 +12,14 @@ export abstract class ITabularConfig {
   clickableRows?: boolean;
   sortBy?: ISortByProperty[];
   cssClass?: string;
-  stickyHeader?: boolean;
   remoteSorting?: boolean;
   multiSorting?: boolean;
+  stickyHeader?: boolean;
+  stickyColumns?: {
+    left: boolean;
+    right: boolean;
+  };
+  theme?: TabularTheme;
 }