+++ src/app/app-routing.module.ts
@@ -23,6 +23,7 @@ import { DateRangePickerComponent } from './core/date-range-picker/date-range-pi
 import {DialogsComponent} from './core/dialogs/dialogs.component';
 import {NgSelectComponent} from './core/ng-select/ng-select.component';
 import {InspectorComponent} from './core/inspector/inspector.component';
+import {LineClampComponent} from './core/line-clamp/line-clamp.component';
 
 const routes: Routes = [
   { path: '', redirectTo: '/install-guide', pathMatch: 'full' },
@@ -48,6 +49,7 @@ const routes: Routes = [
   { path: 'online-status',  component: OnlineStatusComponent },
   { path: 'toastr',  component: ToastrComponent },
   { path: 'ng-select',  component: NgSelectComponent },
+  { path: 'line-clamp',  component: LineClampComponent },
   { path: '**',  component: PageNotFoundComponent },
 ];
 @NgModule({